Modularized the edit and create recipe , made use hooks for them
